requirements.txt
pandas 
python-amazon-sp-api 
matplotlib 